Extracted challenge instructions into their own file
