+++ frontend/src/hacking-instructor/challenges/sqli-1.ts
@@ -1,0 +1,103 @@
+import {
+  waitForInputToHaveValue,
+  waitForInputToNotHaveValue,
+  waitForElementToGetClicked,
+  waitInMs,
+  sleep
+} from '../helpers/helpers'
+import { ChallengeInstruction } from '../'
+
+export const SqlOneInstructions: ChallengeInstruction = {
+  name: 'Login Admin',
+  hints: [
+    {
+      text:
+        "Let's try if we find a way to log in with the administrator's user account. For starters go to the login page.",
+      fixture: '#navbarLoginButton',
+      async resolved () {
+        while (true) {
+          if (window.location.hash === '#/login') {
+            break
+          }
+          await sleep(100)
+        }
+      }
+    },
+    {
+      text: 'To find a way around the normal login protection we will try to use a SQL injection (SQLi).',
+      fixture: '#email',
+      resolved: waitInMs(8000)
+    },
+    {
+      text: "A good starting point for simple SQL injections is to insert quatation marks (like \" or '). These mess with the syntax of the query and might give you indications if a endpoint is vulnarable or not.",
+      fixture: '#email',
+      resolved: waitInMs(15000)
+    },
+    {
+      text: "Start with entering ' in the email field.",
+      fixture: '#email',
+      resolved: waitForInputToHaveValue('#email', "'")
+    },
+    {
+      text: "Now put anything in the password field. Doesn't matter what.",
+      fixture: '#password',
+      resolved: waitForInputToNotHaveValue('#password', '')
+    },
+    {
+      text: 'Press the log in button',
+      fixture: '#loginButton',
+      resolved: waitForElementToGetClicked('#loginButton')
+    },
+    {
+      text: 'Nice! Do you see the error? Maybe you could check the console output ;)',
+      fixture: '#loginButton',
+      resolved: waitInMs(10000)
+    },
+    {
+      text: 'Did you spot the SQL query in the error message? If not, take a look again.',
+      fixture: '#loginButton',
+      resolved: waitInMs(10000)
+    },
+    {
+      text: "Let's try to manipulate the query a bit more. Try out tryping \"' OR true\” into the email field.",
+      fixture: '#email',
+      resolved: waitForInputToHaveValue('#email', "' OR true")
+    },
+    {
+      text: 'Not click the log in button again',
+      fixture: '#loginButton',
+      resolved: waitForElementToGetClicked('#loginButton')
+    },
+    {
+      text: 'Mhh... The query is still failing? Do you see why in the console?',
+      fixture: '#loginButton',
+      resolved: waitInMs(8000)
+    },
+    {
+      text: 'We need to make sure that the rest of the query after our injection doesnt get executed. Any Ideas?',
+      fixture: '#loginButton',
+      resolved: waitInMs(8000)
+    },
+    {
+      text: 'You can comment out the rest of the quries using comments in sql. In sqlite you can use "--" for that',
+      fixture: '#loginButton',
+      resolved: waitInMs(8000)
+    },
+    {
+      text: 'So type in "\' OR true --" in the email field.',
+      fixture: '#email',
+      resolved: waitForInputToHaveValue('#email', "' OR true --")
+    },
+    {
+      text: 'Press the log in button',
+      fixture: '#loginButton',
+      resolved: waitForElementToGetClicked('#login')
+    },
+    {
+      text:
+        'That worked right?! Concratulation on being the new administartor in the shop!.',
+      fixture: '#searchQuery',
+      resolved: waitInMs(10000)
+    }
+  ]
+}
